Refactoring: move Dbms to storage.db package, make credentials access methods public
